refactor(precompiles): rename Zone state module
