+++ crates/precompiles/src/lib.rs
@@ -55,7 +55,6 @@ mod account_keychain;
 pub mod inbox;
 mod nonce;
 pub mod receive_policy_guard;
-pub mod state;
 pub mod storage;
 mod storage_credits;
 pub mod tempo_state;
@@ -63,6 +62,7 @@ pub mod tip403_proxy;
 #[cfg(feature = "std")]
 pub mod tx_context;
 pub mod zone_fee_manager;
+pub mod zone_state;
 pub mod ztip20;
 
 pub use aes_gcm::{AES_GCM_DECRYPT_ADDRESS, AesGcmDecrypt};
